+++ GCL/augmentors/augmentor.py
@@ -1,4 +1,4 @@
-"""Augmentors."""
+"""Augmentors"""
 from __future__ import annotations
 
 import torch
