+++ GCL/augmentors/feature_masking.py
@@ -1,5 +1,5 @@
-from GCL.augmentor.augmentor import PyGGraph, DGLGraph, Augmentor
-from GCL.augmentor.functional import drop_feature
+from GCL.augmentors.augmentor import PyGGraph, DGLGraph, Augmentor
+from GCL.augmentors.functional import drop_feature
 
 
 class FeatureMasking(Augmentor):
