+++ examples/BGRL_L2L.py
@@ -1,8 +1,8 @@
 import copy
 import torch
 import os.path as osp
-import GCL.loss as L
-import GCL.augmentor as A
+import GCL.losses as L
+import GCL.augmentors as A
 import torch.nn.functional as F
 import torch_geometric.transforms as T
 
@@ -10,7 +10,7 @@ from tqdm import tqdm
 from functools import partial
 from torch.optim import Adam
 from GCL.eval import LRTrainableEvaluator, from_PyG_split
-from GCL.model import BootstrapContrast
+from GCL.models import BootstrapContrast
 from sklearn.metrics import f1_score
 from torch_geometric.nn import GCNConv
 from torch_geometric.datasets import WikiCS
