+++ examples/GBT.py
@@ -1,14 +1,14 @@
 import torch
 import os.path as osp
-import GCL.loss as L
-import GCL.augmentor as A
+import GCL.losses as L
+import GCL.augmentors as A
 import torch_geometric.transforms as T
 
 from tqdm import tqdm
 from functools import partial
 from torch.optim import Adam
 from GCL.eval import from_PyG_split, LRTrainableEvaluator
-from GCL.model.contrast_model import WithinEmbedContrast
+from GCL.models.contrast_model import WithinEmbedContrast
 from sklearn.metrics import f1_score
 from torch_geometric.nn import GCNConv
 from torch_geometric.datasets import WikiCS
