+++ examples/MVGRL_graph.py
@@ -1,14 +1,14 @@
 import torch
 import os.path as osp
-import GCL.loss as L
-import GCL.augmentor as A
+import GCL.losses as L
+import GCL.augmentors as A
 
 from torch import nn
 from tqdm import tqdm
 from functools import partial
 from torch.optim import Adam
 from GCL.eval import SVMEvaluator
-from GCL.model import DualBranchContrast
+from GCL.models import DualBranchContrast
 from sklearn.metrics import f1_score
 from sklearn.exceptions import ConvergenceWarning
 from sklearn.utils._testing import ignore_warnings
